requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
numpy
tensorflow-cpu
huggingface_hub
supabase